--- src/main/java/one/stayfocused/backend/service/otp/OtpService.java
@@ -1,8 +1,0 @@
-package one.stayfocused.backend.service;
-
-public interface OtpService {
-
-    String generateOtp(String otpType, String identifier);
-    void validateOtp(String otpType, String identifier, String otp);
-    void deleteOtp(String otpType, String identifier);
-}
