+++ src/main/java/one/stayfocused/backend/service/password/PasswordService.java
@@ -1,4 +1,4 @@
-package one.stayfocused.backend.service;
+package one.stayfocused.backend.service.password;
 
 import one.stayfocused.backend.dto.*;
